Create a dedicated bytes_scanned analyzer rule
