+++ crates/runtime/src/datafusion.rs
@@ -48,8 +48,7 @@ use datafusion::sql::parser::DFParser;
 use datafusion::sql::sqlparser::dialect::PostgreSqlDialect;
 use datafusion::sql::{sqlparser, TableReference};
 use datafusion_federation::{FederatedTableProviderAdaptor, FederationAnalyzerRule};
-use extension::analyzer::SpiceExtensionAnalyzerRule;
-use extension::SpiceQueryPlanner;
+use extension::{bytes_scanned::BytesScannedAnalyzerRule, SpiceQueryPlanner};
 use query::{Protocol, QueryBuilder};
 use snafu::prelude::*;
 use tokio::spawn;
@@ -264,7 +263,7 @@ impl DataFusion {
 
         let ctx = SessionContext::new_with_state(state);
         ctx.add_analyzer_rule(Arc::new(FederationAnalyzerRule::new()));
-        ctx.add_analyzer_rule(Arc::new(SpiceExtensionAnalyzerRule::new()));
+        ctx.add_analyzer_rule(Arc::new(BytesScannedAnalyzerRule::new()));
         ctx.register_udf(embeddings::array_distance::ArrayDistance::new().into());
         ctx.register_udf(crate::datafusion::udf::Greatest::new().into());
         ctx.register_udf(crate::datafusion::udf::Least::new().into());
